api tìm kiếm địa điểm chờ duyệt cho admin

+++ bussiness-service/common/src/main/java/com/example/common/dto/response/PageResponse.java
@@ -12,6 +12,6 @@ import java.util.List;
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class PageResponse<T> {
     List<T> data;
-    int totalPages;
+    long totalPages;
     long totalElements;
 }
